[Feat] 코드 리팩토링, 불필요한 코드 제거
